Basic Fresnel equations framework ready for simulations.

+++ GeneralClasses/Fresnel.h
@@ -15,12 +15,12 @@ public:
 	float rp_wave, tp_wave;  // Fresnel terms for the P-polarized wave
 	float rs_wave, ts_wave;  // Fresnel terms for the S-polarized wave
 
-	void FresnelTermP(int); // Fresnel equations in their general form for the P_polarized wave
-	void FresnelTermS(int); // Fresnel equations in their general form for the S_polarized wave
+	void FresnelTermP(double); // Fresnel equations in their general form for the P_polarized wave
+	void FresnelTermS(double); // Fresnel equations in their general form for the S_polarized wave
 	
 }obj_Fresnel; // Fresnel class object
 
-void Fresnel::FresnelTermP(int theta_i){
+void Fresnel::FresnelTermP(double theta_i){
 	
 	theta_t = asin((obj_Fresnel.n_1/obj_Fresnel.n_2)*sin(theta_i*(M_PI/180)));
 
@@ -32,7 +32,7 @@ void Fresnel::FresnelTermP(int theta_i){
 	obj_Fresnel.tp_wave = 2*(obj_Fresnel.n_1/(obj_Fresnel.u_1*obj_Const.u_0))*cos(theta_i)/(term_1 + term_2);
 }
 
-void Fresnel::FresnelTermS(int theta_i){
+void Fresnel::FresnelTermS(double theta_i){
 	
 	float theta_t = asin((obj_Fresnel.n_1/obj_Fresnel.n_2)*sin(theta_i*(M_PI/180)));
